Feat: Tag 조회 로직 변경 (#10)

+++ src/main/java/com/sw19/sofa/domain/tag/service/TagService.java
@@ -1,25 +1,23 @@
-
 package com.sw19.sofa.domain.tag.service;
 
 import com.sw19.sofa.domain.linkcard.enums.TagType;
 import com.sw19.sofa.domain.tag.entity.Tag;
+import com.sw19.sofa.domain.tag.error.TagErrorCode;
 import com.sw19.sofa.domain.tag.repository.CustomTagRepository;
 import com.sw19.sofa.domain.tag.repository.TagRepository;
 import com.sw19.sofa.global.common.dto.TagDto;
+import com.sw19.sofa.global.error.exception.BusinessException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-
 import java.util.List;
-
 
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class TagService {
     private final TagRepository tagRepository;
-    private final CustomTagRepository customTagRepository;
 
     public List<Tag> getAllTags() {
         return tagRepository.findAll();
@@ -43,14 +41,21 @@ public class TagService {
 
     @Transactional
     protected Tag createTag(String name, TagType type) {
+        if (tagRepository.existsByName(name)) {
+            throw new BusinessException(TagErrorCode.TAG_ALREADY_EXISTS);
+        }
         Tag tag = Tag.builder()
                 .name(name)
                 .type(type)
                 .build();
         return tagRepository.save(tag);
     }
+
     @Transactional
     public void deleteTag(Long id) {
+        if (!tagRepository.existsById(id)) {
+            throw new BusinessException(TagErrorCode.TAG_NOT_FOUND);
+        }
         tagRepository.deleteAllLinkCardTagsByTagId(id);
         tagRepository.deleteById(id);
     }
@@ -61,6 +66,4 @@ public class TagService {
                 .map(TagDto::new)
                 .toList();
     }
-
-
 }
